Fix issue where the amount of divs changed from airport to airport

+++ aipdl.sh
@@ -1,6 +1,9 @@
 #!/bin/bash 
 parameter=$1
 airport_list=()
+ICAO_code=""
+airport_name=""
+
 
 
 
@@ -13,20 +16,29 @@ function get_airport_list() {
     counter=1
     return_string="-"
 
-    wget -U "Mozilla/5.0 (X11; Ubuntu; Linux i686; rv:24.0) Gecko/20100101 Firefox/24.0" -O $local_airport_list $ad_list 
+    echo "local_airport_list=$local_airport_list"
+
+    # We only need to download if it doesnt already exist
+    if [ ! -f $local_airport_list ]; then
+        echo "local airport list not avail, downloading"
+        wget -U "Mozilla/5.0 (X11; Ubuntu; Linux i686; rv:24.0) Gecko/20100101 Firefox/24.0" -O $local_airport_list $ad_list 
+    fi
 
     while :
     do
-        declare -a "start_string=(xmllint --html --xpath \"string(//html/body/div/div/div/div[2]/div[$counter]/h3/a/@href)\" $local_airport_list)"
-        return_string=$(${start_string[@]})
+        params="string(//html/body/div/div/div/div[2]/div[$counter]/h3/a/@href)"
+        airport_object=$(xmllint --html --xpath "$params" $local_airport_list 2> /dev/null)
+        #declare -a "start_string=(xmllint --html --xpath \"string(//html/body/div/div/div/div[2]/div[$counter]/h3/a/@href)\" $local_airport_list)"
+        # return_string=$(${start_string[@]})
         counter=$(( $counter + 1 ))
-        if [ "$return_string" == "" ] 
+        if [ "$airport_object" == "" ] 
         then 
             break
         else
-            airport_list+=($return_string)
+            airport_list+=($airport_object)
         fi
     done
+    echo "*** DEBUG: Finished get_airport_list()"
 }
 
 # Creates URL
@@ -34,11 +46,27 @@ function get_airport_url() {
     echo "*** DEBUG: get_airport_url()"
 
     get_airport_list
+    echo "parameter = $parameter"
     for airport in ${airport_list[@]}; do
-        if [[ "$airport" =~ "$parameter" ]] 
-        then
+        if [[ "$airport" =~ "$parameter" ]]; then
             airport_url="${full_path}EN-AD-2.$1-no-NO.html#AD-2.$1"
-            return 0
+            echo "Found airport url!"
+            echo "airport_url = $airport_url"
+            break
+        fi
+    done
+    echo "*** DEBUG: Finished get_airport_url()"
+}
+
+function ICAO_to_text() {
+    echo "*** DEBUG: ICAO_to_text()"
+
+    get_airport_list
+    for airport in ${airport_list[@]}; do
+        if [[ "$airport" =~ "$ICAO_code" ]]; then
+            echo "ICAO code found!"
+            echo "airport = $airport"
+            break
         fi
     done
 }
@@ -47,20 +75,24 @@ function get_airport_url() {
 function get_chart_meta_data() {
     echo "*** DEBUG: get_chart_meta_data()"
 
+    remote_chart_list="EN-AD-2.$parameter-no-NO.html"
     index=1
     while :
     do
-        chart_name_params="string(//html/body/div[2]/div/div[24]/table/tbody/tr[$index]/td[1]/p)"
+        div_id=$parameter"-AD-2.24"
+        chart_name_params="string(//div[@id='$div_id']/table/tbody/tr[$index]/td[1]/p)"
         chart_name=$(xmllint --html --xpath "$chart_name_params" $remote_chart_list 2> /dev/null)
 
-        chart_code_params="string(//html/body/div[2]/div/div[24]/table/tbody/tr[$index]/td[2]/a)"
+        chart_code_params="string(//div[@id='$div_id']/table/tbody/tr[$index]/td[2]/a)"
         chart_code=$(xmllint --html --xpath "$chart_code_params" $remote_chart_list 2> /dev/null)
 
-        chart_filename_params="string(//html/body/div[2]/div/div[24]/table/tbody/tr[$index]/td[2]/a/@href)"
+        chart_filename_params="string(//div[@id='$div_id']/table/tbody/tr[$index]/td[2]/a/@href)"
         chart_filename=$(xmllint --html --xpath "$chart_filename_params" $remote_chart_list 2> /dev/null)
 
         index=$(( $index + 1 ))
 
+        echo "chart_name = $chart_name"
+        echo "chart_code = $chart_code"
         if [ "$chart_name" == "" ] 
         then
             break
@@ -91,6 +123,7 @@ function download_chart() {
         subdir="Area"
     fi
 
+    # Get substring from the 5th char to the end of ths string
     local_file_name=${chart_filename:5}
     local_chart_name=$(echo $chart_name | awk '{$1=$1;print}')
     local_chart_name="$local_chart_name.pdf"
@@ -103,7 +136,7 @@ function download_chart() {
     fi
     chart_url=$(echo $versioned_url"/"$date"-AIRAC"$local_file_name)
     cd $subdir
-        wget -U "Mozilla/5.0 (X11; Ubuntu; Linux i686; rv:24.0) Gecko/20100101 Firefox/24.0" -O $local_chart_name $chart_url 
+        #wget -U "Mozilla/5.0 (X11; Ubuntu; Linux i686; rv:24.0) Gecko/20100101 Firefox/24.0" -O $local_chart_name $chart_url 
     cd ../..
 
 }
@@ -117,7 +150,7 @@ function download_airport_charts() {
     get_chart_meta_data
 }
 
-
+# This function gets the current version number of the AIS
 function get_current_version() {
     echo "***DEBUG get_current_version()"
     base_url="https://ais.avinor.no/no/AIP/"
@@ -156,22 +189,23 @@ function usage() {
 }
 
 # CHECK FOR COMMAND LINE OPTIONS
-
+get_current_version
 if [[ "$1" == "" ]]; then
     usage
     exit
 elif [[ "$1" == --* ]]; then
     echo "Found option"
     if [ "$1" == "--find" ]; then
-        get_airport_list
+        ICAO_code=$2
+        ICAO_to_text
+        #get_airport_list
     else
         echo "Incorrect syntax"
     fi
 else
-    remote_chart_list="EN-AD-2.$parameter-no-NO.html"
-    get_current_version
+    #get_current_version
     get_meta_files
     download_airport_charts $1
-    rm $remote_chart_list
+    #rm $remote_chart_list
 fi
